% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2k.R
\docType{package}
\name{states2k}
\alias{states2k}
\alias{states2k-package}
\title{states2k: United States of America map with the NAD 1983 Albers
projection}
\description{
states2k provides a map of the USA from the United States Geological
Survey (USGS) with the NAD 1983 Albers projection.
}

