% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2k_info.R
\docType{data}
\name{states2k_info}
\alias{states2k_info}
\title{states2k OGR information}
\format{ogrinfo object}
\usage{
states2k_info
}
\description{
A list containing the OGR information of states2k.
}
\examples{
states2k_info
}

