% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2k_map.R
\docType{data}
\name{states2k_map}
\alias{states2k_map}
\title{states2k_map with OGR read data}
\format{A sp SpatialPolygonsDataFrame with 273 rows and 10 variables:
\describe{
\item{AREA}{Area of region, square meters}
\item{PERIMETER}{Perimeter of region, meters}
\item{STATES2K_}{States2k_}
\item{STATES2K_I}{States2k_I}
\item{STATE}{Integer of states}
\item{NAME}{Name of states}
\item{LSAD}{LSAD}
\item{REGION}{Integer for region}
\item{DIVISION}{Integer for division}
\item{LSAD_TRANS}{LSAD_trans}
}}
\usage{
states2k_map
}
\description{
A spatial polygons data frame containing a map of the USA based on the NAD
1983 Albers projection.
}
\examples{
states2k_map
}

