% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2k_summary.R
\docType{data}
\name{states2k_summary}
\alias{states2k_summary}
\title{states2k Spatial summary}
\format{A summary.Spatial object}
\usage{
states2k_summary
}
\description{
A list containing a spatial summary of states2k.
}
\examples{
states2k_summary
}

