% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statesp010_info.R
\docType{data}
\name{statesp010_info}
\alias{statesp010_info}
\title{statesp010 OGR information}
\format{
ogrinfo object
}
\usage{
statesp010_info
}
\description{
A list containing the OGR information of statesp010.
}
\examples{
statesp010_info
}
