% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statesp010_summary.R
\docType{data}
\name{statesp010_summary}
\alias{statesp010_summary}
\title{statesp010 Spatial summary}
\format{
A summary.Spatial object
}
\usage{
statesp010_summary
}
\description{
A list containing a spatial summary of statesp010.
}
\examples{
statesp010_summary
}
