% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_state_boundaries.R
\docType{package}
\name{USA.state.boundaries}
\alias{USA.state.boundaries}
\title{USA.state.boundaries: 3 sets of United States of America (USA) maps}
\description{
The following maps are included: 1) A map of the USA from the United
States Geological Survey (USGS) with the NAD 1983 Albers projection, 2) A
NAD83 datum map of the USA from the National Map, and 3) A WGS84 datum map
of the USA derived from map 2.
}
