% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_state_boundaries_info.R
\docType{data}
\name{USA_state_boundaries_info}
\alias{USA_state_boundaries_info}
\title{USA_state_boundaries OGR information}
\format{
ogrinfo object
}
\usage{
USA_state_boundaries_info
}
\description{
A list containing the OGR information of USA_state_boundaries.
}
\examples{
USA_state_boundaries_info
}
