% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_state_boundaries_summary.R
\docType{data}
\name{USA_state_boundaries_summary}
\alias{USA_state_boundaries_summary}
\title{USA_state_boundaries Spatial summary}
\format{
A summary.Spatial object
}
\usage{
USA_state_boundaries_summary
}
\description{
A list containing a spatial summary of USA_state_boundaries.
}
\examples{
USA_state_boundaries_summary
}
